refactor: rename LLM provider modules to driver pattern
- Rename llm.rs to llm_drivers.rs and merge with driver_factory.rs
  This consolidates all LLM driver abstractions in one module
- Rename llm_entities.rs to llm_models.rs for clearer naming
  (entities represent LLM models and providers domain)
- Rename provider.rs to driver.rs in everruns-openai and everruns-anthropic
- Rename structs: OpenAIProtocolLlmProvider → OpenAILlmDriver,
  AnthropicLlmProvider → AnthropicLlmDriver
- Update all imports across crates

The driver naming convention better reflects the architectural pattern
where these components translate between generic LLM calls and
provider-specific API protocols.
