+++ crates/everruns-anthropic/src/driver.rs
@@ -1,0 +1,6 @@
+// Anthropic Driver Re-export
+//
+// This module re-exports the AnthropicLlmDriver from everruns-core.
+// The implementation lives in everruns-core to avoid circular dependencies.
+
+pub use everruns_core::anthropic::AnthropicLlmDriver;
