+++ crates/everruns-core/src/lib.rs
@@ -24,7 +24,7 @@ pub mod tool_types;
 pub mod agent;
 pub mod capability_dto;
 pub mod event;
-pub mod llm_entities;
+pub mod llm_models;
 pub mod model_profiles;
 pub mod session;
 pub mod session_file;
@@ -34,7 +34,7 @@ pub mod capabilities;
 pub mod config;
 pub mod error;
 pub mod events;
-pub mod llm;
+pub mod llm_drivers;
 pub mod r#loop;
 pub mod message;
 pub mod step;
@@ -44,9 +44,8 @@ pub mod traits;
 // In-memory implementations for examples and testing
 pub mod memory;
 
-// LLM Protocol providers
+// LLM Driver implementations
 pub mod anthropic;
-pub mod driver_factory;
 pub mod openai;
 
 // Re-exports for convenience
@@ -64,10 +63,11 @@ pub use traits::{
     SessionStore, ToolContext, ToolExecutor,
 };
 
-// LLM types re-exports
-pub use llm::{
-    LlmCallConfig, LlmCallConfigBuilder, LlmCompletionMetadata, LlmContentPart, LlmDriver,
-    LlmMessage, LlmMessageContent, LlmMessageRole, LlmResponse, LlmResponseStream, LlmStreamEvent,
+// LLM driver types re-exports
+pub use llm_drivers::{
+    create_driver, BoxedLlmDriver, LlmCallConfig, LlmCallConfigBuilder, LlmCompletionMetadata,
+    LlmContentPart, LlmDriver, LlmMessage, LlmMessageContent, LlmMessageRole, LlmResponse,
+    LlmResponseStream, LlmStreamEvent, ProviderConfig, ProviderType,
 };
 
 // Tool abstraction re-exports
@@ -95,17 +95,14 @@ pub use atoms::{
 // Tool types (runtime types defined in this crate)
 pub use tool_types::{BuiltinTool, ToolCall, ToolDefinition, ToolPolicy, ToolResult};
 
-// Driver factory re-exports
-pub use driver_factory::{create_driver, BoxedLlmDriver, ProviderConfig, ProviderType};
-
 // Note: CapabilityId and CapabilityStatus are re-exported via capabilities module
 
 // Domain entity re-exports
-// Note: LlmProvider entity is in llm_entities module. Import as: everruns_core::llm_entities::LlmProvider
+// Note: LlmProvider entity is in llm_models module. Import as: everruns_core::llm_models::LlmProvider
 pub use agent::{Agent, AgentStatus};
 pub use capability_dto::{AgentCapability, CapabilityInfo};
 pub use event::Event;
-pub use llm_entities::{
+pub use llm_models::{
     LlmModel, LlmModelCost, LlmModelLimits, LlmModelModalities, LlmModelProfile, LlmModelStatus,
     LlmModelWithProvider, LlmProviderStatus, LlmProviderType, Modality, ReasoningEffort,
     ReasoningEffortConfig, ReasoningEffortValue,
