+++ crates/everruns-core/src/llm_models.rs
