+++ crates/everruns-openai/src/driver.rs
@@ -1,0 +1,6 @@
+// OpenAI Driver Re-export
+//
+// This module re-exports the OpenAILlmDriver from everruns-core.
+// The implementation lives in everruns-core to avoid circular dependencies.
+
+pub use everruns_core::openai::OpenAILlmDriver;
